refactor: rename numericFilters file to match the name of the only exported function in it
