+++ src/libs/algolia/fetchAlgolia/buildNumericFilters.ts
